+++ .gitignore
@@ -1,2 +1,5 @@
 /.env
 /.handler-saves/
+/handler/**
+!/handler/*
+
