--- models/model_div.py
@@ -1,7 +1,0 @@
-def get_div(x, y, i):
-	if i == 1:
-		return int(x) / int(y)
-	if i == 2:
-		return int(x) // int(y)
-	if i == 3:
-		return int(x) % int(y)
