--- models/model_mult.py
@@ -1,4 +1,0 @@
-
-def get_mult(x, y):
-	return int(x) * int(y)
-
