--- models/model_sub.py
@@ -1,3 +1,0 @@
-
-def get_sub(x,y):
-	return int(x) - int(y)
