--- models/model_sum.py
@@ -1,3 +1,0 @@
-
-def get_sum(x,y):
-	return int(x) + int(y)
